cleanup

+++ srcs/utils/get_abs.c
@@ -6,7 +6,7 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 17:04:38 by rbroque           #+#    #+#             */
-/*   Updated: 2022/11/08 17:49:26 by rbroque          ###   ########.fr       */
+/*   Updated: 2022/11/09 11:43:03 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ unsigned int	get_abs(int nb)
 	unsigned int	abs;
 
 	if (nb < 0)
-		abs = -nb;
+		abs = -(unsigned int)nb;
 	else
 		abs = nb;
 	return (abs);
